+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+- Add convenience method for obtaining elfi samples as `InferenceData`` to be used with `arviz`
 - Improve `randmaxvar` batch acquisitions and initialisation by enabling sampling from prior
 - Drop official Python support for 3.7 and 3.8 as GPy is not officially supported for these versions
 - Enable using `maxiter` in `bo.utils.minimize`